+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Format JSON command responses as key: value lines
+
 ## [0.2.0] - 2026-04-03
 
 ### Added